# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.updateMeRequest:
    properties:
      avatar_url:
        type: string
      banner_url:
        type: string
      bio:
        type: string
      username:
        type: string
    type: object
...
  model.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      pillar:
        description: '"dev" | "ai"'
        type: string
      slug:
        type: string
    type: object
...
  model.ContentStatus:
    enum:
    - pending_moderation
    - approved
    - flagged
    - blocked
    type: string
    x-enum-varnames:
    - StatusPendingModeration
    - StatusApproved
    - StatusFlagged
    - StatusBlocked
  model.Post:
    properties:
      author:
        allOf:
        - $ref: '#/definitions/model.User'
        description: Relations
      category:
        $ref: '#/definitions/model.Category'
      category_id:
        type: string
      comment_count:
        type: integer
      content:
        description: TipTap HTML
        type: string
      created_at:
        type: string
      id:
        type: string
      images:
        items:
          $ref: '#/definitions/model.PostImage'
        type: array
      is_hidden:
        type: boolean
      status:
        $ref: '#/definitions/model.ContentStatus'
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      user_vote:
        type: integer
      vote_count:
        description: Champs calculés (non stockés en DB)
        type: integer
    type: object
  model.PostImage:
    properties:
      created_at:
        type: string
      id:
        type: string
      mime_type:
        type: string
      post_id:
        type: string
      size_bytes:
        type: integer
      url:
        type: string
    type: object
  model.User:
    properties:
      avatar_url:
        type: string
      banner_url:
        type: string
      bio:
        type: string
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      role:
        $ref: '#/definitions/model.UserRole'
      updated_at:
        type: string
      username:
        type: string
    type: object
  model.UserRole:
    enum:
    - user
    - moderator
    - admin
    type: string
    x-enum-varnames:
    - RoleUser
    - RoleModerator
    - RoleAdmin
...
  service.PublicProfile:
    properties:
      follower_count:
        type: integer
      following_count:
        type: integer
      is_following:
        type: boolean
      post_count:
        type: integer
      recent_posts:
        items:
          $ref: '#/definitions/model.Post'
        type: array
      user:
        $ref: '#/definitions/model.User'
    type: object
...
host: localhost:8080
info:
  contact:
    name: DevHelp Team
    url: https://github.com/bloiss/DevHelp
  description: API REST du forum DevHelp — entraide Dev & IA.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: DevHelp API
  version: "1.0"
paths:
...
  /api/v1/auth/github/callback:
    get:
      description: Reçoit le code d'autorisation GitHub, échange les tokens, crée
        ou connecte l'utilisateur et redirige vers le frontend avec les tokens JWT.
      parameters:
      - description: State CSRF
        in: query
        name: state
        required: true
        type: string
      - description: Code d'autorisation
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "307":
          description: Redirect to /auth/login?error=oauth_failed
          schema:
            type: string
      summary: Callback GitHub OAuth
      tags:
      - auth
...
  /api/v1/auth/google/callback:
    get:
      description: Reçoit le code d'autorisation Google, échange les tokens, crée
        ou connecte l'utilisateur et redirige vers le frontend avec les tokens JWT.
      parameters:
      - description: State CSRF
        in: query
        name: state
        required: true
        type: string
      - description: Code d'autorisation
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "307":
          description: Redirect to /auth/login?error=oauth_failed
          schema:
            type: string
      summary: Callback Google OAuth
      tags:
      - auth
...
  /api/v1/users/{username}:
    get:
      description: Retourne le profil public d'un utilisateur avec ses statistiques
        (posts, followers, following).
      parameters:
      - description: Nom d'utilisateur
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.PublicProfile'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Profil public d'un utilisateur
      tags:
      - users
...
  /api/v1/users/me:
    patch:
      consumes:
      - application/json
      description: Met à jour les informations du profil de l'utilisateur connecté
        (username, avatar, bannière, bio).
      parameters:
      - description: Champs à modifier
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.updateMeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mettre à jour son propre profil
      tags:
      - users
  /api/v1/users/search:
    get:
      description: Recherche des utilisateurs dont le username correspond à la requête.
        Utile pour l'autocomplétion.
      parameters:
      - description: Terme de recherche (min 1 caractère)
        in: query
        name: q
        required: true
        type: string
      - description: Nombre max de résultats (défaut 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Rechercher des utilisateurs par username
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Entrez "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"